+++ src/Module.php
@@ -399,8 +399,16 @@ class Module extends \yii\base\Module implements BootstrapInterface
         echo $view->renderDynamic('return Yii::$app->getModule("' . $this->getUniqueId() . '")->getToolbarHtml();');
 
         // echo is used in order to support cases where asset manager is not available
-        echo '<style>' . $view->renderPhpFile(__DIR__ . '/assets/css/toolbar.css') . '</style>';
-        echo '<script>' . $view->renderPhpFile(__DIR__ . '/assets/js/toolbar.js') . '</script>';
+        if ($view->getAssetManager()) {
+            $css = $view->renderPhpFile(__DIR__ . '/assets/css/toolbar.css');
+            echo $view->registerCss($css);
+
+            $js = $view->renderPhpFile(__DIR__ . '/assets/js/toolbar.js');
+            echo $view->registerJs($js);
+        } else {
+            echo '<style>' . $view->renderPhpFile(__DIR__ . '/assets/css/toolbar.css') . '</style>';
+            echo '<script>' . $view->renderPhpFile(__DIR__ . '/assets/js/toolbar.js') . '</script>';
+        }
     }
 
     /**